artist modal add

+++ app/src/main/java/cloud/ptl/boardgamecollector/db/entity/Artist.java
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Artist {
     @PrimaryKey(autoGenerate = true)
-    public int artistId;
+    public Long artistId;
 
     public String name;
     public String surname;
